Implement Public API

+++ traders-api/src/main/java/com/jdc/one/traders/model/dto/output/CategoryDto.java
@@ -3,7 +3,7 @@ package com.jdc.one.traders.model.dto.output;
 public record CategoryDto(
 		int id, 
 		String name,
-		int products
+		long products
 		) {
 
 }
