+++ traders-api/src/main/java/com/jdc/one/traders/model/dto/output/TopSellerDto.java
@@ -3,7 +3,7 @@ package com.jdc.one.traders.model.dto.output;
 public record TopSellerDto(
 		int id,
 		String name,
-		int sales
+		long sales
 		) {
 
 }
